Extract out ProblemNavButton for reusability
