+++ frontend/src/components/core/Button.tsx
@@ -220,3 +220,28 @@ export const InvertedSmallButton = styled(SmallButton)`
   color: ${({ theme }) => theme.colors.text};
   background: ${({ theme }) => theme.colors.white};
 `;
+
+type ProblemNavButtonProps = {
+  disabled: boolean,
+};
+
+export const ProblemNavButton = styled(DefaultButton)<ProblemNavButtonProps>`
+  font-size: ${({ theme }) => theme.fontSize.default};
+  color: ${({ theme }) => theme.colors.gray};
+  background-color: ${({ theme, disabled }) => (disabled ? theme.colors.background : theme.colors.white)};
+  border-radius: 5px;
+  width: 35px;
+  height: 35px;
+  margin: 5px;
+  
+  box-shadow: 0 1px 6px rgba(0, 0, 0, 0.16);
+  
+  &:hover {
+    box-shadow: ${({ disabled }) => (disabled ? '0 1px 6px rgba(0, 0, 0, 0.16)' : '0 1px 6px rgba(0, 0, 0, 0.20)')};
+    cursor: ${({ disabled }) => (disabled ? 'default' : 'pointer')}; 
+  }
+  
+  i {
+    line-height: 35px;
+  }
+`;
