+++ frontend/src/components/game/ProblemPanel.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import MarkdownEditor from 'rich-markdown-editor';
 import { BottomFooterText, ProblemHeaderText, SmallText } from '../core/Text';
-import { DefaultButton, getDifficultyDisplayButton } from '../core/Button';
+import { getDifficultyDisplayButton, ProblemNavButton } from '../core/Button';
 import { Copyable } from '../special/CopyIndicator';
 import {
   CenteredContainer,
@@ -38,31 +38,6 @@ const ProblemNavContainer = styled(FlexRight)`
 
 const ProblemCountText = styled(SmallText)`
   color: gray;
-`;
-
-type ProblemNavButtonProps = {
-  disabled: boolean,
-};
-
-const ProblemNavButton = styled(DefaultButton)<ProblemNavButtonProps>`
-  font-size: ${({ theme }) => theme.fontSize.default};
-  color: ${({ theme }) => theme.colors.gray};
-  background-color: ${({ theme, disabled }) => (disabled ? theme.colors.background : theme.colors.white)};
-  border-radius: 5px;
-  width: 35px;
-  height: 35px;
-  margin: 5px;
-  
-  box-shadow: 0 1px 6px rgba(0, 0, 0, 0.16);
-  
-  &:hover {
-    box-shadow: ${({ disabled }) => (disabled ? '0 1px 6px rgba(0, 0, 0, 0.16)' : '0 1px 6px rgba(0, 0, 0, 0.20)')};
-    cursor: ${({ disabled }) => (disabled ? 'default' : 'pointer')}; 
-  }
-  
-  i {
-    line-height: 35px;
-  }
 `;
 
 type ProblemPanelProps = {
